chore: split migration in 2
