+++ src/graphql/error-map.ts
@@ -449,7 +449,6 @@ export const mapError = (error: ApplicationError): CustomApolloError => {
     case "MissingExpiredAtKratosError":
     case "MissingTotpKratosError":
     case "IncompatibleSchemaUpgradeError":
-    case "UnconfirmedOnChainTxError":
     case "InvalidDeviceTokenError":
       message = `Unexpected error occurred, please try again or contact support if it persists (code: ${
         error.name
